+++ libraries/bio/src/utils/cell-renderer-monomer-placer.ts
@@ -12,16 +12,9 @@ import {SeqSplittedBase} from './macromolecule/types';
 
 type MonomerPlacerProps = {
   seqHandler: SeqHandler,
-  monomerLib?: IMonomerLib,
+  monomerLib: IMonomerLib,
   monomerCharWidth: number, separatorWidth: number,
   monomerToShort: MonomerToShortFunc, monomerLengthLimit: number,
-};
-
-const polymerTypeMap = {
-  [ALPHABET.DNA]: HELM_POLYMER_TYPE.RNA,
-  [ALPHABET.RNA]: HELM_POLYMER_TYPE.RNA,
-  [ALPHABET.PT]: HELM_POLYMER_TYPE.PEPTIDE,
-  [ALPHABET.UN]: HELM_POLYMER_TYPE.PEPTIDE,
 };
 
 export class MonomerPlacer {
@@ -29,14 +22,14 @@ export class MonomerPlacer {
 
   // width of separator symbol
   private separatorWidth = 5;
-  private props: MonomerPlacerProps;
+  public props: MonomerPlacerProps;
   private _rowsProcessed: DG.BitSet; // rows for which monomer lengths were processed
 
   private _updated: boolean = false;
   public get updated(): boolean { return this._updated; }
 
   public _monomerLengthMap: { [key: string]: TextMetrics } = {}; // caches the lengths to save time on g.measureText
-  public _monomerStructureMap: { [key: string]: HTMLDivElement } = {}; // caches the atomic structures of monomers
+  public _monomerStructureMap: { [key: string]: HTMLElement } = {}; // caches the atomic structures of monomers
   private readonly subs: Unsubscribable[] = [];
 
   /** View is required to subscribe and handle for data frame changes */
@@ -54,15 +47,15 @@ export class MonomerPlacer {
           this.props = this.propsProvider();
           this._monomerLengthList = null;
           this._rowsProcessed = DG.BitSet.create(this.col.length);
-        } catch (e) {
-          console.error(e);
+        } catch (err: any) {
+          console.error(err);
         }
       }));
       this.subs.push(grok.events.onViewRemoved.subscribe((view: DG.View) => {
         try {
           if (this.grid?.view?.id === view.id) this.destroy();
-        } catch (e) {
-          console.error(e);
+        } catch (err) {
+          console.error(err);
         }
       }));
     }
@@ -179,12 +172,6 @@ export class MonomerPlacer {
       }
     }
     return left;
-  }
-
-  public getMonomer(symbol: string): Monomer | null {
-    const alphabet = this.props.seqHandler.alphabet ?? ALPHABET.UN;
-    const polymerType = polymerTypeMap[alphabet as ALPHABET];
-    return this.props.monomerLib?.getMonomer(polymerType, symbol) ?? null;
   }
 
   public setMonomerLengthLimit(limit: number): void {