+++ libraries/ml/CHANGELOG.md
@@ -1,5 +1,13 @@
 # ml changelog
 
+## 6.6.0 (2024-04-14)
+
+### Features
+
+* Add webGPU Sparse matrix calculation.
+* Add webGPU UMAP implementation.
+* Add webGPU option for activity cliffs + MCL
+
 ## 6.5.1 (2024-04-05)
 
 Fix function editor for seq/chem space